Use int for opcodes to prevent sign issues, add size impl

+++ api/src/main/java/com/intel/i40eaqdebug/api/logs/LogEntry.java
@@ -9,7 +9,7 @@ public interface LogEntry {
     TimeStamp getTimeStamp();
     int getErr();
     short getFlags();
-    short getOpCode();
+    int getOpCode();
     short getRetVal();
     int getCookieHigh();
     int getCookieLow();
